FROMLIST: sched/fair: jump to max OPP when crossing UP threshold
Since the true utilization of a long running task is not detectable while
it is running and might be bigger than the current cpu capacity, create the
maximum cpu capacity head room by requesting the maximum cpu capacity once
the cpu usage plus the capacity margin exceeds the current capacity. This
is also done to try to harm the performance of a task the least.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Juri Lelli <juri.lelli@arm.com>

(am from https://patchwork.kernel.org/patch/6738071)
Signed-off-by: Juri Lelli <juri.lelli@arm.com>

Change-Id: Ic7cded62657d5932adb202a6946d5cccd99ea4bb